pyproject.toml
[tool.poetry]
name = "jax-dependency"
version = "0.1.0"
description = ""
authors = ["Daniel <ddold@htwg-konstanz.de>"]
readme = "README.md"
packages = [{include = "jax_dependency"}]

[tool.poetry.dependencies]
python = "^3.10"
ipywidgets = "^8.1.3"
widgetsnbextension = "^4.0.11" 
flax = "^0.8.5"
#jaxlib = {version = ">=0.4.30,<0.5.0", extras = ["cuda12.cudnn89"], source = "jaxlib"}
jax = {version = "0.4.30", extras = ["cuda12"], source = "jax"}
seaborn = "^0.13.2"
ipykernel = "^6.29.5"
numpyro = "^0.15.0"
pandas = "^2.2.2"
arviz = "^0.18.0"
matplotlib = "^3.9.0"
blackjax = "^1.2.1"
notebook = "^7.2.1"
graphviz = "^0.20.3"
hvplot = "^0.10.0"
datashader = "^0.16.3"
funsor = "^0.4.5"
wandb = "^0.17.8"
jax-tqdm = "^0.2.2"
torch = {version = "^2.4.0", source = "torch_cuda"}
torchvision = {version = "^0.19.0", source = "torch_cuda"}
nvidia-cuda-nvcc-cu12 = ">=12.4.0, <12.6.0"
jaxopt = "^0.8.3"
scikit-learn = "^1.5.2"
tabulate = "^0.9.0"
regex = "^2024.11.6"


[[tool.poetry.source]]
name = "jax"
url = "https://storage.googleapis.com/jax-releases/jax_releases.html"
priority = "explicit"


[[tool.poetry.source]]
name = "jaxlib"
url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
priority = "explicit"


[[tool.poetry.source]]
name = "torch_cuda"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[[tool.poetry.source]]
name = "torch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
